requirements.txt
streamlit
transformers
torch
requests
pillow
